+++ src/main/java/nz/ac/auckland/se306/group12/models/ScheduledTask.java
@@ -12,7 +12,7 @@ import lombok.Getter;
 public class ScheduledTask {
 
   private final Node node;
-  private final long startTime;
-  private final long processor;
+  private final int startTime;
+  private final int processor;
 
 }
